+++ crates/ryo3-core/src/lib.rs
@@ -13,8 +13,8 @@ pub use pystring::{
 };
 pub use pytryfrom::PyTryFrom;
 pub use ryo3_macro_rules::{
-    py_io_err, py_io_error, py_key_err, py_not_implemented_err, py_not_implemented_error,
-    py_runtime_err, py_runtime_error, py_type_err, py_type_error, py_value_err, py_value_error,
-    pytodo,
+    any_repr, py_io_err, py_io_error, py_key_err, py_not_implemented_err, py_not_implemented_error,
+    py_overflow_err, py_overflow_error, py_runtime_err, py_runtime_error, py_type_err,
+    py_type_error, py_value_err, py_value_error, pytodo,
 };
 pub use sync::{PyLock, PyRead, PyWrite, RyMutex, RyRwLock};