+++ crates/ryo3-uuid/src/py_uuid.rs
@@ -5,8 +5,8 @@ use pyo3::prelude::*;
 use pyo3::sync::PyOnceLock;
 use pyo3::types::PyTuple;
 use pyo3::{BoundObject, intern};
-use ryo3_bytes::PyBytes;
-use ryo3_macro_rules::{any_repr, py_type_err, py_value_err, py_value_error, pytodo};
+use ryo3_bytes::ExactReadableBuffer;
+use ryo3_core::{PyAsciiString, any_repr, py_type_err, py_value_err, py_value_error, pytodo};
 
 static NODE_CACHE: OnceLock<u64> = OnceLock::new();
 
@@ -80,8 +80,8 @@ impl PyUuid {
     )]
     fn py_new(
         hex: Option<&str>,
-        bytes: Option<PyBytes>,
-        bytes_le: Option<PyBytes>,
+        bytes: Option<PyUuidBytes>,
+        bytes_le: Option<PyUuidBytes>,
         fields: Option<&Bound<PyTuple>>,
         int: Option<u128>,
         version: Option<u8>,
@@ -94,8 +94,8 @@ impl PyUuid {
 
         let py_uuid = match (hex, bytes, bytes_le, fields, int) {
             (Some(hex), None, None, None, None) => Self::from_hex(hex),
-            (None, Some(bytes), None, None, None) => Self::from_pybytes(bytes),
-            (None, None, Some(bytes_le), None, None) => Self::from_bytes_le(bytes_le),
+            (None, Some(bytes), None, None, None) => Ok(Self::from_pybytes(bytes)),
+            (None, None, Some(bytes_le), None, None) => Ok(Self::from_bytes_le(bytes_le)),
             (None, None, None, Some(fields), None) => Self::from_fields(fields),
             (None, None, None, None, Some(int)) => Ok(Self::from_int(int)),
             // emsg taken from the python itself
@@ -138,7 +138,8 @@ impl PyUuid {
     }
 
     fn __getnewargs__<'py>(&self, py: Python<'py>) -> PyResult<Bound<'py, PyTuple>> {
-        PyTuple::new(py, vec![self.0.hyphenated().to_string()])
+        let py_ascii_str: PyAsciiString = self.0.hyphenated().to_string().into();
+        PyTuple::new(py, vec![py_ascii_str])
     }
 
     #[cfg(any(target_pointer_width = "32", target_pointer_width = "64"))]
@@ -161,16 +162,16 @@ impl PyUuid {
     }
 
     #[pyo3(name = "to_string")]
-    fn py_to_string(&self) -> String {
+    fn py_to_string(&self) -> PyAsciiString {
         self.__str__()
     }
 
-    fn __str__(&self) -> String {
-        self.0.to_string()
+    fn __str__(&self) -> PyAsciiString {
+        self.0.to_string().into()
     }
 
-    fn __repr__(&self) -> String {
-        format!("UUID('{}')", self.py_to_string())
+    fn __repr__(&self) -> PyAsciiString {
+        format!("{self}").into()
     }
 
     fn __int__(&self) -> u128 {
@@ -242,19 +243,13 @@ impl PyUuid {
 
     #[staticmethod]
     #[pyo3(name = "from_bytes")]
-    #[expect(clippy::needless_pass_by_value)]
-    fn from_pybytes(b: PyBytes) -> PyResult<Self> {
-        uuid::Uuid::from_slice(b.as_ref())
-            .map(PyUuid)
-            .map_err(|e| py_value_error!("invalid UUID bytes: {}", e))
+    fn from_pybytes(b: PyUuidBytes) -> Self {
+        Self(uuid::Uuid::from_bytes(b.0))
     }
 
     #[staticmethod]
-    #[expect(clippy::needless_pass_by_value)]
-    fn from_bytes_le(b: PyBytes) -> PyResult<Self> {
-        uuid::Uuid::from_slice_le(b.as_ref())
-            .map(PyUuid)
-            .map_err(|e| py_value_error!("invalid UUID bytes_le: {}", e))
+    fn from_bytes_le(b: PyUuidBytes) -> Self {
+        Self(uuid::Uuid::from_bytes_le(b.0))
     }
 
     // | Field                      | Meaning                          |
@@ -283,8 +278,7 @@ impl PyUuid {
     }
 
     fn __bytes__<'py>(&self, py: Python<'py>) -> Bound<'py, pyo3::types::PyBytes> {
-        let bytes = self.0.as_bytes().to_vec();
-        pyo3::types::PyBytes::new(py, &bytes)
+        pyo3::types::PyBytes::new(py, self.0.as_bytes())
     }
 
     #[getter]
@@ -294,7 +288,7 @@ impl PyUuid {
 
     #[getter]
     fn bytes_le<'py>(&self, py: Python<'py>) -> Bound<'py, pyo3::types::PyBytes> {
-        let bytes = self.0.to_bytes_le().to_vec();
+        let bytes = self.0.to_bytes_le();
         pyo3::types::PyBytes::new(py, &bytes)
     }
 
@@ -323,8 +317,8 @@ impl PyUuid {
     }
 
     #[getter]
-    fn hex(&self) -> String {
-        self.0.simple().to_string()
+    fn hex(&self) -> PyAsciiString {
+        self.0.simple().to_string().into()
     }
 
     #[getter]
@@ -398,7 +392,6 @@ impl PyUuid {
         } else if let Ok(s) = value.extract::<&str>() {
             Self::from_str(s).map(|dt| dt.into_pyobject(py))?
         } else if let Ok(b) = value.extract::<[u8; 16]>() {
-            // let s = String::from_utf8_lossy(pybytes.as_bytes());
             Self::from_int(u128::from_be_bytes(b)).into_pyobject(py)
         } else if let Ok(pybytes) = value.cast::<pyo3::types::PyBytes>() {
             let s = String::from_utf8_lossy(pybytes.as_bytes());
@@ -420,7 +413,7 @@ impl PyUuid {
         value: &Bound<'py, PyAny>,
         _handler: &Bound<'py, PyAny>,
     ) -> PyResult<Bound<'py, Self>> {
-        use ryo3_macro_rules::py_value_error;
+        use ryo3_core::py_value_error;
         Self::from_any(value).map_err(|e| py_value_error!("UUID validation error: {e}"))
     }
 
@@ -433,6 +426,12 @@ impl PyUuid {
     ) -> PyResult<Bound<'py, PyAny>> {
         use ryo3_pydantic::GetPydanticCoreSchemaCls;
         Self::get_pydantic_core_schema(cls, source, handler)
+    }
+}
+
+impl std::fmt::Display for PyUuid {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "UUID('{}')", self.0)
     }
 }
 
@@ -568,14 +567,12 @@ pub fn uuid7(timestamp: Option<u64>) -> PyResult<PyUuid> {
     }
 }
 
-#[pyfunction(
-    signature = (a = None, b = None, c = None, *, buf = None),
-)]
+#[pyfunction(signature = (a = None, b = None, c = None, *, buf = None))]
 pub fn uuid8(
     a: Option<u64>,
     b: Option<u16>,
     c: Option<u64>,
-    buf: Option<PyBytes>,
+    buf: Option<PyUuidBytes>,
 ) -> PyResult<PyUuid> {
     use rand::prelude::*;
 
@@ -586,12 +583,7 @@ pub fn uuid8(
                 return py_value_err!("uuid8(): pass either bytes=... or a/b/c, not both",);
             }
         }
-        // extract the bytes as [u8; 16]
-        let slice: &[u8; 16] = bts
-            .as_slice()
-            .try_into()
-            .map_err(|_| py_value_error!("uuid8(bytes=...): bytes must be exactly 16 bytes",))?;
-        return Ok(PyUuid::from(uuid::Uuid::new_v8(*slice)));
+        return Ok(PyUuid::from(uuid::Uuid::new_v8(bts.0)));
     }
 
     let mut rng = rand::rng();
@@ -710,5 +702,20 @@ impl ryo3_pydantic::GetPydanticCoreSchemaCls for PyUuid {
             args,
             Some(&serialization_kwargs),
         )
+    }
+}
+
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub struct PyUuidBytes([u8; 16]);
+
+impl<'a, 'py> FromPyObject<'a, 'py> for PyUuidBytes {
+    type Error = PyErr;
+
+    fn extract(ob: Borrowed<'a, 'py, PyAny>) -> PyResult<Self> {
+        if let Ok(b) = ob.extract::<ExactReadableBuffer<'a, 'py, 16>>() {
+            Ok(Self(*b.as_array()))
+        } else {
+            py_type_err!("Expected a bytes-like object with length == 16")
+        }
     }
 }